refactor defineEnvironment to also accept HTML and MathML builders, fixes #870
